v0.0.14 - refactor and relocate options input
