+++ src/configuration/subgraph/entities/service_entity_field/exclude_from_input/mod.rs
@@ -1,0 +1,13 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+pub enum ExcludeFromInput {
+    FindOne,
+    FindMany,
+    CreateOne,
+    UpdateOne,
+    UpdateMany,
+    UpdateOneQuery,
+    UpdateManyQuery,
+    All,
+}
