+++ src/data_sources/mongo/services/find_one/mod.rs
@@ -7,7 +7,7 @@ use mongodb::{
 
 use crate::{
     data_sources::mongo::{EagerLoadOptions, MongoDataSource},
-    graphql::schema::OptionsInput,
+    graphql::schema::create_options_input::OptionsInput,
 };
 
 use super::Services;
