+++ src/data_sources/sql/create_query/create_find_many_query/mod.rs
@@ -8,7 +8,7 @@ use crate::{
         data_sources::sql::DialectEnum, entities::ServiceEntityConfig, SubGraphConfig,
     },
     data_sources::sql::{SqlDataSource, SqlValueEnum},
-    graphql::schema::{DirectionEnum, SortInput},
+    graphql::schema::create_options_input::{DirectionEnum, SortInput},
 };
 
 use super::{create_nested_query_recursive::FilterOperator, JoinClauses};
