+++ src/graphql/schema/create_options_input/mod.rs
@@ -1,6 +1,71 @@
+use std::fmt::Display;
+
 use async_graphql::dynamic::{Enum, EnumItem, InputObject, InputValue, TypeRef};
+use serde::{Deserialize, Serialize};
 
 use crate::graphql::schema::ServiceSchema;
+
+#[derive(Clone, Debug, Serialize, PartialEq)]
+pub enum DirectionEnum {
+    Asc,
+    Desc,
+}
+
+//Implement disaplay
+impl Display for DirectionEnum {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            DirectionEnum::Asc => write!(f, "Asc"),
+            DirectionEnum::Desc => write!(f, "Desc"),
+        }
+    }
+}
+
+impl std::str::FromStr for DirectionEnum {
+    type Err = async_graphql::Error;
+
+    fn from_str(input: &str) -> Result<DirectionEnum, Self::Err> {
+        match input {
+            "ASC" | "Asc" => Ok(DirectionEnum::Asc),
+            "Desc" | "DESC" => Ok(DirectionEnum::Desc),
+            _ => Err(async_graphql::Error::new(format!(
+                "Invalid DirectionEnum: {}",
+                input
+            ))),
+        }
+    }
+}
+
+// Accept ASC Asc DESC Desc - Deserializer for DirectionEnum
+impl<'de> serde::Deserialize<'de> for DirectionEnum {
+    fn deserialize<D>(deserializer: D) -> Result<DirectionEnum, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        match s.as_str() {
+            "ASC" | "Asc" => Ok(DirectionEnum::Asc),
+            "Desc" | "DESC" => Ok(DirectionEnum::Desc),
+            _ => Err(serde::de::Error::custom(format!(
+                "Invalid DirectionEnum: {}",
+                s
+            ))),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+pub struct SortInput {
+    pub field: String,
+    pub direction: DirectionEnum,
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+pub struct OptionsInput {
+    pub per_page: Option<i32>,
+    pub page: Option<i32>,
+    pub sort: Option<Vec<SortInput>>,
+}
 
 impl ServiceSchema {
     pub fn create_options_input(mut self) -> Self {
